EFA-392 Move methods to LocalMessage.kt
Make method private in MessageViewViewModel.
