+++ k9mail/src/main/java/com/fsck/k9/extensions/Message.kt
@@ -8,13 +8,3 @@ import com.fsck.k9.planck.PlanckUtils
 
 fun Message.getRatingFromHeader() =
     getHeader(MimeHeader.HEADER_PEP_RATING).firstOrNull()?.let { PlanckUtils.stringToRating(it) }
-
-fun LocalMessage?.isValidForHandshake() = this != null
-        && (account?.isPlanckPrivacyProtected ?: false)
-        && from != null
-        && from.size == 1
-        && PlanckUtils.isRatingReliable(planckRating)
-        && getRecipients(RecipientType.CC).isNullOrEmpty()
-        && getRecipients(RecipientType.BCC).isNullOrEmpty()
-        && !getRecipients(RecipientType.TO).isNullOrEmpty()
-        && !from.first().address.equals(account.email, ignoreCase = true) // sender not my own account
